refactor: :truck: Move the getPublicKeysWithBackoffRetry functionality to getpublickeys.go as makes more sense there
