+++ internal/azurejwtvalidator/getpublickeys.go
@@ -1,6 +1,7 @@
 package azurejwtvalidator
 
 import (
+	"context"
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/base64"
@@ -12,6 +13,7 @@ import (
 	"net/http"
 	"strings"
 
+	"github.com/cenkalti/backoff/v5"
 	"github.com/music-tribe/azadjwtvalidation/internal/jwtmodels"
 )
 
@@ -116,5 +118,16 @@ func (azjwt *AzureJwtValidator) verifyAndSetPublicKey(publicKey string) error {
 		azjwt.rsakeys.Set(kid, pubKey)
 	}
 
+	return nil
+}
+
+func (azjwt *AzureJwtValidator) getPublicKeysWithBackoffRetry(ctx context.Context, maxRetries uint) error {
+	operation := func() (string, error) {
+		return "", azjwt.GetPublicKeys()
+	}
+	_, err := backoff.Retry(ctx, operation, backoff.WithMaxTries(maxRetries), backoff.WithBackOff(backoff.NewExponentialBackOff()))
+	if err != nil {
+		return err
+	}
 	return nil
 }
